remove helpers package and move contents to util package
