+++ pkg/iter/sample_iterator.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"io"
 
-	"github.com/grafana/loki/pkg/helpers"
 	"github.com/grafana/loki/pkg/logproto"
 	"github.com/grafana/loki/pkg/logql/stats"
 	"github.com/grafana/loki/pkg/util"
@@ -195,7 +194,7 @@ func (i *heapSampleIterator) requeue(ei SampleIterator, advanced bool) {
 	if err := ei.Error(); err != nil {
 		i.errs = append(i.errs, err)
 	}
-	helpers.LogError("closing iterator", ei.Close)
+	util.LogError("closing iterator", ei.Close)
 }
 
 type sampletuple struct {
