+++ pkg/util/math.go
@@ -1,4 +1,4 @@
-package helpers
+package util
 
 // MinUint32 return the min of a and b.
 func MinUint32(a, b uint32) uint32 {
